некоторые поля в db string, instagram (all)

+++ database/migrations/2018_11_07_095550_create_authors_table.php
@@ -15,7 +15,7 @@ class CreateAuthorsTable extends Migration
     {
         Schema::create('authors', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('author_id');
+            $table->bigInteger('author_id');
             $table->string('first_name');
             $table->string('last_name');
             $table->string('avatar');
